Extract swift_static_framework rule out of mixed_static_framework.
This is to improve the debugging experience of the Bazel based Xcode
project.

Currently we have the `swift_static_framework` rule as just a macro for
`mixed_static_framework` rule, mostly because they share many things in
the implementation details. However (probably) due to the many hacks in
the `mixed_static_framework` rule, some of debugger commands became
unusable when using the Bazel based Xcode project -- for instance, `po`
and `expr`. This patch extracts `swift_static_framework` out to a real
rule in an attempt to fix that.

The first culprit I found is the underlying Obj-C module map that we
generate for compiling the Swift part of the module -- although it
compiles fine, but when attaching the binary to the debugger,
`ClangImporter` still complains about redefinitions of the same module.
For pure Swift modules, that module map is unnecessary -- it's generated
but doesn't include any header -- so by extracting the
`swift_static_framework` rule out, we were able to remove that part.

This patch makes the above debugger commands work if you build something
that doesn't have any `mixed_static_framework` target in your build's
dependency graph.