Support custom avoid_deps
This patch adds a new `avoid_deps` attribute to
`{objc,swift,mixed}_static_framework` rules. This attribute is a list of
`objc_library` and `swift_library` targets on which the framework
depends in order to compile, but the transitive closure of which will
not be linked into the framework's binary. By default this is the same
as `deps`, that is none of the depependencies will be linked into the
framework's binary. For example, providing an empty list (`[]`) here
will result in a fully static link binary.

Fixes https://github.com/line/bazel_rules_apple/issues/5.